+++ Dependiject/Store.swift
@@ -67,7 +67,7 @@ public struct Store<ObjectType> {
     
     /// Create a stored value on a custom scheduler.
     ///
-    /// Use this init to schedule updates on a specific scheduler other than `RunLoop.main`.
+    /// Use this init to schedule updates on a specific scheduler other than `DispatchQueue.main`.
     public init<S: Scheduler>(
         wrappedValue: ObjectType,
         on scheduler: S,
@@ -92,7 +92,7 @@ public struct Store<ObjectType> {
     ///
     /// To control when updates are published, see ``init(wrappedValue:on:schedulerOptions:)``.
     public init(wrappedValue: ObjectType) {
-        self.init(wrappedValue: wrappedValue, on: RunLoop.main)
+        self.init(wrappedValue: wrappedValue, on: DispatchQueue.main)
     }
     
     /// An equivalent to SwiftUI's