+++ CHANGELOG.md
@@ -5,6 +5,8 @@ formatting guidelines.
 
 ## Unreleased
 
+## [1.0.0] - 2022-12-13
+
 ### Removed
 
 - Dropped support for CocoaPods &lt; 1.11.3.
@@ -75,3 +77,4 @@ Initial release.
 [0.1.0]: https://github.com/Tiny-Home-Consulting/Dependiject/tree/0.1.0
 [0.2.0]: https://github.com/Tiny-Home-Consulting/Dependiject/tree/0.2.0
 [0.3.0]: https://github.com/Tiny-Home-Consulting/Dependiject/tree/0.3.0
+[1.0.0]: https://github.com/Tiny-Home-Consulting/Dependiject/tree/1.0.0